Revert "Homework #4"
