+++ angular.json
@@ -17,11 +17,6 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "stylePreprocessorOptions": {
-              "includePaths": [
-                "src/assets/styles"
-              ]
-            },
             "outputPath": "dist/todo",
             "index": "src/index.html",
             "main": "src/main.ts",
@@ -87,11 +82,6 @@
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "stylePreprocessorOptions": {
-              "includePaths": [
-                "src/assets/styles"
-              ]
-            },
             "polyfills": [
               "zone.js",
               "zone.js/testing"
