+++ src/app/app.component.html
@@ -1,3 +1,1 @@
 <app-todo-list></app-todo-list>
-
-<notification-container></notification-container>
