+++ src/app/app.module.ts
@@ -7,23 +7,18 @@ import { TodoListComponent } from './components/todo-list/todo-list.component';
 import { TodoItemComponent } from './components/todo-item/todo-item.component';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { SharedModule } from './components/shared/shared.module';
+import { SharedModule } from './components/shared/shared/shared.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatInputModule } from '@angular/material/input';
-import { TodoDescriptionComponent } from './components/todo-description/todo-description.component';
-import { NotificationService } from './components/shared/notification/notification-message.service';
 
 @NgModule({
   declarations: [
     AppComponent,
     TodoListComponent,
     TodoItemComponent,
-    TodoDescriptionComponent,
   ],
   imports: [
-    // Angular
     BrowserModule,
-    BrowserAnimationsModule,
     AppRoutingModule,
     CommonModule,
     FormsModule,
@@ -33,10 +28,9 @@ import { NotificationService } from './components/shared/notification/notificati
 
     // Custom
     SharedModule,
+      BrowserAnimationsModule,
   ],
-  providers: [
-    { provide: NotificationService, useClass: NotificationService }
-  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
