+++ src/app/components/shared/button/button.component.html
@@ -1,3 +1,1 @@
-<button type="button" [disabled]="disabled">
-    <ng-content></ng-content>
-</button>
+<button type="button" >{{ title }}</button>
