+++ src/app/components/shared/button/button.component.scss
@@ -1,0 +1,25 @@
+@import '../../../../variables.scss';
+
+// .button {
+//     // background-color: $primary-color;
+//     // color: white;
+//     // border: none;
+    
+//     transition: .5s all;
+//     height: 100%;
+//     border-radius: 5px;
+//     white-space: nowrap;
+//     padding: 5px 10px;
+
+//     &.button--add {
+//         background-color: $success-color;
+//     }
+
+//     &.button--delete {
+//         background-color: $danger-color;
+//     }
+
+//     &.button:disabled {
+//         opacity: .5;
+//     }
+// }
