+++ src/app/components/shared/button/button.component.ts
@@ -7,6 +7,6 @@ import { Component, Input } from '@angular/core';
 })
 export class ButtonComponent {
 
-  @Input('disabled')
-  disabled: boolean = false;
+  @Input({ required: true })
+  title: string = '';
 }
