--- src/app/components/shared/notification/notification-container/notification-container.component.html
@@ -1,1 +1,0 @@
-<notification-message></notification-message>
