--- src/app/components/shared/notification/notification-container/notification-container.component.scss
@@ -1,11 +1,0 @@
-:host {
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-
-    position: fixed; 
-    bottom: 40px;
-    right: 40px;
-    
-    z-index: 999999999;
-}
