--- src/app/components/shared/notification/notification-container/notification-container.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'notification-container',
-  templateUrl: './notification-container.component.html',
-  styleUrls: ['./notification-container.component.scss'],
-})
-export class NotificationContainerComponent { }
