--- src/app/components/shared/notification/notification-message/notification-message.component.scss
@@ -1,110 +1,0 @@
-@import 'images-variables.scss';
-
-.i-base {
-    background-size: cover;
-    background-repeat: no-repeat;
-
-    mask-repeat: no-repeat;
-    -webkit-mask-repeat: no-repeat;
-    
-    &.i-close {
-        background-color: #4e5260;
-        
-        mask-size: 14px 14px;
-        mask-image: url($close-svg);
-        -webkit-mask-image: url($close-svg);
-    }
-
-    &.i-close:hover {
-        cursor: pointer;
-        background-color: #000000;
-    }
-
-    &.i-success {
-        background-image: url($success-svg);
-    }
-
-    &.i-danger {
-        background-image: url($danger-svg);
-    }
-}
-
-.notification {
-    position: relative;
-    width: 420px;
-
-    margin-top: 20px;
-
-    background: #FFFFFF;
-    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-    border-radius: 7px;
-
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-}
-
-.notification__header {
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    padding: 20px 18px;
-    border-radius: 7px 7px 0px 0px;
-    max-height: 32px;
-    
-    * {
-        font-weight: 500;
-        font-size: 1rem;
-        color: #4B4B4B;
-    }
-}
-
-.notification__body {
-    font-size: 14px;
-    color: #4B4B4B;
-    border-radius: 7px;
-}
-
-.header__icon {
-    width: 18px;
-    height: 18px;
-}
-
-.header__content {
-    display: flex;
-    gap: 10px;
-    flex: 1;
-}
-
-.header__close {
-    width: 16px;
-    height: 16px;
-
-    position: relative;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.body__content {
-    padding: 20px;
-}
-
-.notification__header--success {
-    background-color: rgba(19, 180, 203, 0.2);
-
-    .header__icon {
-        background-size: 18px 18px;
-        background-image: url($success-svg);
-    }
-}
-
-.notification__header--danger {
-    background-color: #FCC9C9;
-
-    .header__icon {
-        background-size: 18px 18px;
-        background-image: url($danger-svg);
-    }
-}
