+++ src/app/components/shared/shared/shared.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { ButtonComponent } from '../button/button.component';
+
+@NgModule({
+  declarations: [
+    ButtonComponent,
+  ],
+  imports: [
+    CommonModule,
+  ],
+  exports: [
+    ButtonComponent,
+  ],
+})
+export class SharedModule { }
