--- src/app/components/shared/tooltip/container/tooltip.component.html
@@ -1,1 +1,0 @@
-<span>{{ title }}</span>
