--- src/app/components/shared/tooltip/container/tooltip.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { TooltipComponent } from './tooltip.component';
-
-describe('TooltipComponent', () => {
-  let component: TooltipComponent;
-  let fixture: ComponentFixture<TooltipComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [TooltipComponent]
-    });
-    fixture = TestBed.createComponent(TooltipComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
