--- src/app/components/shared/tooltip/container/tooltip.interface.ts
@@ -1,5 +1,0 @@
-export interface ITooltip {
-    title: string;
-    left: number;
-    top: number;
-}
