--- src/app/components/shared/tooltip/container/tooltip.variables.scss
@@ -1,13 +1,0 @@
-// tooltip
-$tooltip-margin-top: calc(7px + 3px); // Отступ 3px.
-$tooltip-padding: 3px 6px;
-$tooltip-border-radius: 5px;
-$tooltip-font-weight: 500;
-
-// common
-$tooltip-font-size: .825rem;
-$backgroud-color: rgb(173, 157, 255);
-$font-color: rgb(23, 0, 27);
-$background-color: #17a2b8;
-$box-shadow: 4px 3px 6px -1px rgba(0, 0, 0, 0.3);
-$z-index: 9999;
