--- src/app/components/shared/tooltip/directive/tooltip.directive.spec.ts
@@ -1,8 +1,0 @@
-import { TooltipDirective } from './tooltip.directive';
-
-describe('TooltipDirective', () => {
-  it('should create an instance', () => {
-    // const directive = new TooltipDirective();
-    // expect(directive).toBeTruthy();
-  });
-});
