--- src/app/components/todo-description/todo-description.component.html
@@ -1,1 +1,0 @@
-<p>{{ description }}</p>
