--- src/app/components/todo-description/todo-description.component.scss
@@ -1,12 +1,0 @@
-:host {
-    animation: showItem 500ms 1 ease-in;
-}
-
-@keyframes showItem {
-    0% {
-        opacity: 0;
-    }
-    100% {
-        opacity: 1;
-    }
-}
