+++ src/app/components/todo-item/todo-item.component.scss
@@ -8,27 +8,6 @@ $background-item-color: #ededed;
     min-height: 48px;
     padding: 0 10px;
     border-radius: 5px;
-    opacity: .5;
+    
     background-color: $background-item-color;
-    
-    transition: all .3s ease-in;
-
-    &:focus-within {
-        box-shadow: 0px 0px 0px 3px #0052ff26;
-    }
-    
-    &:hover {
-        opacity: 1;     
-        background-color: $background-item-color;
-        cursor: pointer;
-    }
-
-    &.item--active {
-        opacity: 1;
-    }
-}
-
-.change-title-container {
-    display: flex;
-    gap: 10px;
 }
