+++ src/app/components/todo-list/todo-list.component.scss
@@ -28,7 +28,6 @@ ul {
     padding-left: 0;
 
     li {
-        position: relative;
         list-style-type: none;
     }
 }
