+++ src/app/components/todo-list/todo.data.ts
@@ -3,23 +3,19 @@ import { ITodo, ITodos } from "../../interface/todo/todo-item.interface";
 export const TODOS: ITodos = [
     {
         id: 1,
-        text: 'Buy a new gaming laptop',
-        description: 'Buy a new gaming laptop description'
+        text: 'Buy a new gaming laptop'
     },
     {
         id: 2,
-        text: 'Complete previous task',
-        description: 'Complete previous task description'
+        text: 'Complete previous task'
     },
     {
         id: 3,
-        text: 'Create some Angular app',
-        description: 'Create some Angular app description'    
+        text: 'Create some Angular app'
     }
 ];
 
 export const INITIAL_TODO: ITodo = {
     id: -1,
-    text: '',
-    description: ''
+    text: ''
 }
