+++ src/app/interface/todo/todo-item.interface.ts
@@ -1,7 +1,6 @@
 export interface ITodo {
     id: number;
     text: string;
-    description: string;
 };
 
 export type ITodos = Array<ITodo>;
