--- src/assets/styles/images-variables.scss
@@ -1,10 +1,0 @@
-$images-path: '~src/assets/images';
-
-$close-svg: $images-path + '/close.svg';
-$close-round-svg: $images-path + '/close-round.svg';
-
-$success-svg: $images-path  + '/success.svg';
-$success-round-svg: $images-path  + '/success-round.svg';
-
-$danger-svg: $images-path  + '/danger.svg';
-$danger-round-svg: $images-path  + '/danger-round.svg';
