+++ src/styles.scss
@@ -44,7 +44,7 @@ body {
     font-family: Roboto, "Helvetica Neue", sans-serif; 
 }
 
-@import './assets/styles/color-variables.scss';
+@import './variables.scss';
 
 // Spinner 
 @keyframes spinner-border {
