+++ src/variables.scss
@@ -1,4 +1,3 @@
-  // Colors
   $danger-color: #ff4444;
   $warning-color: #ffbb33;
   $success-color: #00c851;
